requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
python-whois>=0.8.0
spacy>=3.7.0
nltk>=3.8.0
Flask>=3.0.0
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
redis>=5.0.1
rq>=1.15.1
